Integrated calculate_qc_metrics into gen_qc/run_qc
This function computes a few of the quantities we already care
about and a bunch more that we _might_ care about.  So using this
moving forward.

Also added a list of hemoglobin genes for the three primary genomes
so that we can use similar loading functionality instead of relying
on Hbb all the time.