Replaced filter_genes_dispersion with highly_variable_genes
This changes the workflow a bit so that there is no distinction
between "adata" and "adata_filt".  Highly variable genes are now
indiciated via a column in adata.var and pretty much the only thing
that we need to tell to look there is PCA, which is done.
Everything else uses PCs, so I think we're good.